include input value when iterating over entries in autocomplete
When iterating over entries in autocomplete dropdown, when you are
about to wrap over (pressing down arrow would move you to top entry
and vice versa with up arrow), no entry is being focused and input is
populated with default (old) value. This version is in alpha state and
has to be cleaned and tested.