Save original input value (#1)
* include package-lock.json

* include input value when iterating over entries in autocomplete

When iterating over entries in autocomplete dropdown, when you are
about to wrap over (pressing down arrow would move you to top entry
and vice versa with up arrow), no entry is being focused and input is
populated with default (old) value. This version is in alpha state and
has to be cleaned and tested.

* change active status of entry which is currently active, return the rest unchanged

* change state's var name to be more descriptive

* add tests for saving original input value and disabling focus when pressing arrow down/up key while on last/first item

* test if onChange function for input is called with appropriate values

* add props for enabling including input value when iterating over autocomplete

Apart from that clean tests a bit, by moving spy further up the scope,
and resetting it state before each test case. That way we don't have
shallowly render wrapper again in spy test case just to include the spy.

* add prop for clearing input when pressing escape button

* add test case for clearing/resetting/restoring input value when pressing escape button

* move test autocomplete data to global const

* rename prop for setting input to its original value after clearing autocomplete

Include test for this prop in test block for `includeInputValue` since
it depends on value from state which saved only when
`includeInputValue` prop flag is true.

* add first draft fo documentation of prop features

* make prop names more descriptive

Also try to add better description for props.

* rename local state entry name

* update package.json; don't try to deploy demo

* remove lock file